requirements.txt
Django>=5.2.4
djangorestframework>=3.15.1
PyJWT>=2.8.0
gunicorn>=22.0.0
python-dotenv>=1.0.1
psycopg2-binary>=2.9.9  # If you plan to use PostgreSQL
